+++ game_patch/rf/os/console.h
@@ -85,4 +85,5 @@ namespace rf::console
     static auto& history_max_index = addr_as_ref<int>(0x005A4030);
     static auto& console_is_visible = addr_as_ref<bool()>(0x0050B520);
     static auto& console_keep_history = addr_as_ref<bool>(0x005A402C);
+    static auto& background_color = addr_as_ref<Color>(0x017751DC);
 }